+++ packages/hoc-react/src/wrapComponent.tsx
@@ -1,9 +1,11 @@
 import * as React from 'react'
-import { useContext, useEffect } from 'react'
+import { useContext, useEffect, useState } from 'react'
 import { withRouter, RouteComponentProps } from 'react-router-dom'
-import { FC, Action } from 'ssr-types-react'
+import { FC, IFC, Action, IWindow } from 'ssr-types-react'
 
-let routerChanged = false
+declare const window: IWindow
+
+let hasRender = false
 
 const fetch = async (WrappedComponent: FC, dispatch: React.Dispatch<Action>, props: RouteComponentProps) => {
   let asyncLayoutData = {}
@@ -24,27 +26,32 @@ const fetch = async (WrappedComponent: FC, dispatch: React.Dispatch<Action>, pro
   })
 }
 
-function wrapComponent (WrappedComponent: FC) {
+function wrapComponent (WrappedComponent: FC|IFC) {
   return withRouter(props => {
+    const [ready, setReady] = useState(WrappedComponent.name !== 'dynamicComponent')
     const { dispatch } = useContext(window.STORE_CONTEXT)
-
     useEffect(() => {
       didMount()
     }, [])
-
     const didMount = async () => {
-      if (routerChanged || !window.__USE_SSR__) {
+      if (hasRender || !window.__USE_SSR__) {
         // ssr 情况下只有路由切换的时候才需要调用 fetch
         // csr 情况首次访问页面也需要调用 fetch
-        await fetch(WrappedComponent, dispatch, props)
+        await fetch((WrappedComponent as FC), dispatch, props)
+        if (WrappedComponent.name === 'dynamicComponent') {
+          WrappedComponent = (await (WrappedComponent as FC)()).default
+          setReady(true)
+        }
       }
-      if (!routerChanged) {
-        // routerChanged 为 true 代表已经进行过切换路由的操作
-        routerChanged = true
+      if (!hasRender) {
+        // ssr 场景首次渲染的情况下客户端无需获取数据
+        hasRender = true
       }
     }
-
-    return <WrappedComponent {...props}></WrappedComponent>
+    return (
+      // @ts-expect-error
+      ready ? <WrappedComponent {...props}></WrappedComponent> : null
+    )
   })
 }