+++ packages/plugin-react/src/entry/client-entry.tsx
@@ -5,7 +5,7 @@ import { preloadComponent } from 'ssr-client-utils'
 import { wrapComponent } from 'ssr-hoc-react'
 import { IWindow, LayoutProps, ReactClientESMFeRouteItem, ReactClientRoutesType } from 'ssr-types-react'
 // @ts-expect-error
-import * as Routes from 'ssr-temporary-routes'
+import * as Routes from '_build/ssr-temporary-routes'
 import { AppContext } from './context'
 
 const { FeRoutes, layoutFetch, App, BASE_NAME } = Routes as ReactClientRoutesType
@@ -19,7 +19,6 @@ const clientRender = async (): Promise<void> => {
   }
   // 客户端渲染||hydrate
   const routes = await preloadComponent(FeRoutes, BASE_NAME)
-
   ReactDOM[window.__USE_SSR__ ? 'hydrate' : 'render'](
     <BrowserRouter basename={BASE_NAME}>
       <AppContext>
@@ -47,5 +46,12 @@ const clientRender = async (): Promise<void> => {
     module?.hot?.accept?.() // webpack 场景下的 hmr
   }
 }
+if (!window.__disableClientRender__) {
+  // 如果服务端直出的时候带上该记号，则默认不进行客户端渲染，将处理逻辑交给上层
+  // 可用于微前端场景下自定义什么时候进行组件渲染的逻辑调用
+  clientRender()
+}
 
-export default clientRender()
+export {
+  clientRender
+}