+++ packages/plugin-vue/src/entry/client-entry.ts
@@ -3,7 +3,7 @@ import { Store } from 'vuex'
 import { Route } from 'vue-router'
 import { findRoute, normalizePath } from 'ssr-client-utils'
 // @ts-expect-error
-import * as Routes from 'ssr-temporary-routes'
+import * as Routes from '_build/ssr-temporary-routes'
 import { ESMFetch, RoutesType, IClientFeRouteItem } from './interface'
 import { createRouter, createStore } from './create'
 
@@ -77,4 +77,12 @@ const clientRender = async () => {
   }
 }
 
-export default clientRender()
+if (!window.__disableClientRender__) {
+  // 如果服务端直出的时候带上该记号，则默认不进行客户端渲染，将处理逻辑交给上层
+  // 可用于微前端场景下自定义什么时候进行组件渲染的逻辑调用
+  clientRender()
+}
+
+export {
+  clientRender
+}