+++ packages/types-react/src/route.ts
@@ -7,6 +7,7 @@ export interface LayoutProps {
   config?: IConfig
   children?: JSX.Element
   staticList?: StaticList
+  injectState?: any
 }
 export interface StaticList {
   injectCss: JSX.Element[]
@@ -28,16 +29,22 @@ export type ReactESMFetch = () => Promise<{
 
 export type ESMLayout = () => Promise<React.FC<LayoutProps>>
 
-export interface FC<T={}> extends React.FC<T> {
+export interface IFC<T={}> extends React.FC<T> {
+
+}
+
+export interface FC<T = {}> {
+  (): Promise<{
+    default: IFC<T>
+  }>
   fetch?: ReactESMFetch
   layoutFetch?: ReactFetch
-  preload?: () => Promise<FC>
 }
 
 export type ReactServerESMFeRouteItem<T = {}, U={}> = {
   path: string
   fetch?: ReactFetch
-  component: FC<T>
+  component: IFC<T>
   webpackChunkName: string
 } & U
 
@@ -54,6 +61,8 @@ export interface ReactRoutesType {
   layoutFetch: ReactFetch
   FeRoutes: ReactServerESMFeRouteItem[]
   BASE_NAME?: string
+  state?: any
+  reducer?: any
 }
 export interface ReactClientRoutesType {
   Layout: React.FC<LayoutProps>
@@ -61,6 +70,8 @@ export interface ReactClientRoutesType {
   layoutFetch: ReactFetch
   FeRoutes: ReactClientESMFeRouteItem[]
   BASE_NAME?: string
+  state?: any
+  reducer?: any
 }
 
 export interface IContext<T=any> {