+++ packages/plugin-nestjs/CHANGELOG.md
@@ -3,6 +3,285 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.6.14](https://github.com/ykfe/ssr/compare/v5.6.13...v5.6.14) (2021-08-25)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.6.13](https://github.com/ykfe/ssr/compare/v5.6.12...v5.6.13) (2021-08-25)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.6.12](https://github.com/ykfe/ssr/compare/v5.6.11...v5.6.12) (2021-08-25)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.6.10](https://github.com/ykfe/ssr/compare/v5.6.10-alpha.0...v5.6.10) (2021-08-24)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.6.9](https://github.com/ykfe/ssr/compare/v5.6.8...v5.6.9) (2021-08-24)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.6.8](https://github.com/ykfe/ssr/compare/v5.6.7...v5.6.8) (2021-08-24)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.6.7](https://github.com/ykfe/ssr/compare/v5.6.6...v5.6.7) (2021-08-24)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.6.6](https://github.com/ykfe/ssr/compare/v5.6.5...v5.6.6) (2021-08-23)
+
+
+### Bug Fixes
+
+* lint ([1a647a7](https://github.com/ykfe/ssr/commit/1a647a73756fdb69c6cde0188e62ed782317e524))
+
+
+
+
+
+## [5.6.4](https://github.com/ykfe/ssr/compare/v5.6.3...v5.6.4) (2021-08-13)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.6.3](https://github.com/ykfe/ssr/compare/v5.6.2...v5.6.3) (2021-08-12)
+
+
+### Bug Fixes
+
+* 类型修复 nestjs 新增自定义提示 ([d17bf51](https://github.com/ykfe/ssr/commit/d17bf51e6a313a2f57abd87fde7ec014cf724da1))
+
+
+
+
+
+## [5.6.2](https://github.com/ykfe/ssr/compare/v5.6.1...v5.6.2) (2021-08-11)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.99](https://github.com/ykfe/ssr/compare/v5.5.98...v5.5.99) (2021-08-09)
+
+
+### Bug Fixes
+
+* stdout color force in child_process ([b6e5666](https://github.com/ykfe/ssr/commit/b6e5666aaa8aa45923b5f9f9bbc1cabef94fc244))
+
+
+
+
+
+## [5.5.98](https://github.com/ykfe/ssr/compare/v5.5.97...v5.5.98) (2021-08-06)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.97](https://github.com/ykfe/ssr/compare/v5.5.96...v5.5.97) (2021-08-06)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.96](https://github.com/ykfe/ssr/compare/v5.5.94...v5.5.96) (2021-08-05)
+
+
+### Bug Fixes
+
+* ssr types ([742440f](https://github.com/ykfe/ssr/commit/742440f186af3577d0df0bf5fac25d665e43dbcd))
+
+
+
+
+
+## [5.5.95](https://github.com/ykfe/ssr/compare/v5.5.94...v5.5.95) (2021-08-05)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.94](https://github.com/ykfe/ssr/compare/v5.5.93...v5.5.94) (2021-08-05)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.92](https://github.com/ykfe/ssr/compare/v5.5.91...v5.5.92) (2021-08-03)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.91](https://github.com/ykfe/ssr/compare/v5.5.90...v5.5.91) (2021-08-02)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.90](https://github.com/ykfe/ssr/compare/v5.5.89...v5.5.90) (2021-08-02)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.89](https://github.com/ykfe/ssr/compare/v5.5.88...v5.5.89) (2021-07-31)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.88](https://github.com/ykfe/ssr/compare/v5.5.87...v5.5.88) (2021-07-30)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.87](https://github.com/ykfe/ssr/compare/v5.5.86...v5.5.87) (2021-07-30)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.84](https://github.com/ykfe/ssr/compare/v5.5.83...v5.5.84) (2021-07-26)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.83](https://github.com/ykfe/ssr/compare/v5.5.82...v5.5.83) (2021-07-26)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.80](https://github.com/ykfe/ssr/compare/v5.5.79...v5.5.80) (2021-07-25)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.79](https://github.com/ykfe/ssr/compare/v5.5.78...v5.5.79) (2021-07-24)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.77](https://github.com/ykfe/ssr/compare/v5.5.76...v5.5.77) (2021-07-22)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.76](https://github.com/ykfe/ssr/compare/v5.5.75...v5.5.76) (2021-07-21)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.75](https://github.com/ykfe/ssr/compare/v5.5.74...v5.5.75) (2021-07-19)
+
+
+### Bug Fixes
+
+* @types/node use peerdependencies ([93d3cbd](https://github.com/ykfe/ssr/commit/93d3cbd6fe0851062c09965b9572dedb0f1ce0a5))
+
+
+
+
+
+## [5.5.74](https://github.com/ykfe/ssr/compare/v5.5.73...v5.5.74) (2021-07-16)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.73](https://github.com/ykfe/ssr/compare/v5.5.73-alpha.3...v5.5.73) (2021-07-16)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
+## [5.5.72](https://github.com/ykfe/ssr/compare/v5.5.71...v5.5.72) (2021-07-15)
+
+**Note:** Version bump only for package ssr-plugin-nestjs
+
+
+
+
+
 ## [5.5.71](https://github.com/ykfe/ssr/compare/v5.5.70...v5.5.71) (2021-07-12)
 
 **Note:** Version bump only for package ssr-plugin-nestjs